Remove architecture-specific Argo CD binaries after installation in Vagrantfile

+++ Vagrantfile
@@ -171,7 +171,8 @@ EOF1
     [ $(uname -m) = x86_64 ] && curl -sSL -o argocd-linux-amd64 https://github.com/argoproj/argo-cd/releases/latest/download/argocd-linux-amd64 && sudo install -m 555 argocd-linux-amd64 /usr/local/bin/argocd
     [ $(uname -m) = aarch64 ] && curl -sSL -o argocd-linux-arm64 https://github.com/argoproj/argo-cd/releases/latest/download/argocd-linux-arm64 && sudo install -m 555 argocd-linux-arm64 /usr/local/bin/argocd
     sudo install -m 555 argocd-linux-amd64 /usr/local/bin/argocd
-    rm argocd-linux-amd64
+    [ $(uname -m) = x86_64 ] && rm argocd-linux-amd64
+    [ $(uname -m) = aarch64 ] && rm argocd-linux-arm64
 
     kubectl create namespace argocd
